refactor (webutils): Move common aggregation builder routine to webutils
