+++ applications/concept-cat/pom.xml
@@ -86,7 +86,7 @@
         <dependency>
             <groupId>no.dcat</groupId>
             <artifactId>webutils-spring</artifactId>
-            <version>1.1.0</version>
+            <version>1.2.0</version>
         </dependency>
         <dependency>
             <groupId>org.apache.jena</groupId>
