+++ applications/concept-cat/src/main/java/no/ccat/controller/ConceptSearchController.java
@@ -2,13 +2,12 @@ package no.ccat.controller;
 
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import no.ccat.hateoas.PagedResourceWithAggregations;
 import no.ccat.model.ConceptDenormalized;
+import no.fdk.webutils.aggregation.ResponseUtil;
 import org.elasticsearch.index.query.BoolQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.search.aggregations.AbstractAggregationBuilder;
-import org.elasticsearch.search.aggregations.Aggregation;
 import org.elasticsearch.search.aggregations.AggregationBuilders;
 import org.elasticsearch.search.aggregations.bucket.terms.Terms;
 import org.slf4j.Logger;
@@ -27,10 +26,7 @@ import org.springframework.hateoas.PagedResources;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @CrossOrigin
 @RestController
@@ -129,7 +125,7 @@ public class ConceptSearchController {
         }
 
         if (!StringUtils.isEmpty(sortfield)) {
-                addSort(sortfield, sortdirection, finalQuery);
+            addSort(sortfield, sortdirection, finalQuery);
         }
 
         AggregatedPage<ConceptDenormalized> aggregatedPage = elasticsearchTemplate.queryForPage(finalQuery, ConceptDenormalized.class);
@@ -151,37 +147,6 @@ public class ConceptSearchController {
         }
     }
 
-    static class QueryUtil {
-        static QueryBuilder createTermFilter(String term, String value) {
-            return value.equals(MISSING) ?
-                QueryBuilders.boolQuery().mustNot(QueryBuilders.existsQuery(term)) :
-                QueryBuilders.termQuery(term, value);
-        }
-    }
-
-    static class ResponseUtil {
-        static PagedResourceWithAggregations<ConceptDenormalized> addAggregations(PagedResources<ConceptDenormalized> conceptResources, AggregatedPage<ConceptDenormalized> aggregatedPage) {
-            return new PagedResourceWithAggregations<>(conceptResources, extractAggregations(aggregatedPage));
-        }
-
-        static Map<String, no.ccat.model.Aggregation> extractAggregations(AggregatedPage<ConceptDenormalized> aggregatedPage) {
-            Map<String, Aggregation> aggregationsElastic = aggregatedPage.getAggregations().getAsMap();
-            //todo this aggregating seems now to be an emergent pattern, shared with api, refacor out as common (together with the harvest history)
-            final Map<String, no.ccat.model.Aggregation> aggregations = new HashMap<>();
-            aggregationsElastic.forEach((aggregationName, aggregation) -> {
-                no.ccat.model.Aggregation outputAggregation = new no.ccat.model.Aggregation() {{
-                    buckets = new ArrayList<>();
-                }};
-
-                ((Terms) aggregation).getBuckets().forEach((bucket) -> {
-                    outputAggregation.getBuckets().add(no.ccat.model.AggregationBucket.of(bucket.getKeyAsString(), bucket.getDocCount()));
-                });
-                aggregations.put(aggregationName, outputAggregation);
-            });
-            return aggregations;
-        }
-    }
-
     private void addSort(String sortfield, String sortdirection, NativeSearchQuery searchBuilder) {
         if (!sortfield.trim().isEmpty()) {
 
@@ -198,6 +163,14 @@ public class ConceptSearchController {
 
             logger.debug("sort: {}", sort.toString());
             searchBuilder.addSort(sort);
+        }
+    }
+
+    static class QueryUtil {
+        static QueryBuilder createTermFilter(String term, String value) {
+            return value.equals(MISSING) ?
+                QueryBuilders.boolQuery().mustNot(QueryBuilders.existsQuery(term)) :
+                QueryBuilders.termQuery(term, value);
         }
     }
 }
