+++ applications/search-api/pom.xml
@@ -123,7 +123,7 @@
         <dependency>
             <groupId>no.dcat</groupId>
             <artifactId>webutils-spring</artifactId>
-            <version>1.1.0</version>
+            <version>1.2.0</version>
         </dependency>
 
         <dependency>
